Refactor ProjectStructureValidator -> ChangeCollector
